Update Follow Interactive Marker
TeleopCallBack checks if received pose is empty, if not fills the geometry_msgs::Pose and sets has_new_msg boolean as true.

In main loop while(ros:ok())

If new message arrived sending IK solver a request, and waiting for the response. When response received as joint positions fills the joint_group_positions vector to be send as command.

When no new message arrives (interactive marker does not move) last position of the interactive marker will be converted to joint space and sent as command.